+++ Core/Scene/ScaleManipulator.cpp
@@ -463,11 +463,11 @@ void ScaleManipulator::MouseMove( const MouseMoveInput& e )
 
     // Pick ray from our starting location
     Line startRay;
-    m_View->GetCamera()->ViewportToLine(sX, sY, startRay);
+    m_View->GetCamera()->ViewportToLine( (f32)sX, (f32)sY, startRay);
 
     // Pick ray from our current location
     Line endRay;
-    m_View->GetCamera()->ViewportToLine(eX, eY, endRay);
+    m_View->GetCamera()->ViewportToLine( (f32)eX, (f32)eY, endRay);
 
     // start and end points of the drag in world space, on the line or on the plane
     Vector3 p1, p2;
